🚚 move CardReader into src/
